block scope and functional scope
This code will result in a ReferenceError.

The reason for this is that when using the `let` keyword to declare the variable `a` within the block `if (true)`, the scope of `a` is limited to that block. As a result, attempting to access `a` outside of the block will result in a `ReferenceError`.

If the `let` keyword were changed to `var`, the code would run without an error because `var` has function scope rather than block scope. In that case, the output would be `2`