Q10
Use this URL - https://example.com/api/profile to make a fake fetch call and handle errors if any. Show a proper message to the user on the DOM, as per the status and message received from the server. A fakeFetch has been provided. Use HTML, CSS & JS template in REPL or Vanilla template in CodeSandbox for this question

Output on the DOM should be:
Unauthorized Access! Looks like you are not logged in. Please login to see your profile data.