Q2
Use this URL - https://example.com/winner-team to make a fake fetch call to the get names of winner group members and show a nice message on the DOM congratulating all of them. Use HTML, CSS & JS template in REPL or Vanilla template in CodeSandbox for this question.

Output on the DOM should be:
Congratulation to the members of winning team Jhon, Raju, Anjali, Sakshi, great work keep it up.