Q10
Use this URL - https://example.com/getImage to make a fake fetch call which takes a url and dimensions for the photo to be displayed. Dimensions should be passed in the format [width, height]. Show the image provided in the response on DOM. Use HTML, CSS & JS template in REPL or Vanilla template in CodeSandbox for this question.

Dimensions can be 200/300/100. For example you can pass [200, 200] or [200, 300], etc.

Output on the DOM should be an image.