Q2
Write an ES6 function checkForTeam which takes a obj with properties firstname , lastname and a optional property team . It will check if a student already has a team. If team is not given then add them to team “A” and return the object else do nothing and return the same object. Avoid using in-built methods.