Q4
Given an array of todoItems

a. Build a React component that takes the todoItems as props and list all the todo tasks.

b. if isDone is true then show the item with a strikethrough.

c. Add a X button against each item and remove that task from the list on click of the button.