Q5
Given an array of todoItems

a. Build a React component that takes the todoItems as props and list all the todo tasks.

b. Add a "Mark as Done" button against each item and toggle isDone property value between true and false on click of the button. If isDone is true change the color of that task to red with a strikethrough and if it is false remove the strikethrough and the red color.