Q7
Create a React component called Comments.

a. Fetch the comments using thefake fetch and list the data on DOM.

b. Each comment component will have the text, user’s name and a delete button.

c. On click of the delete button, that particular comment object should be deleted and should not be visible on the DOM.