Q4
Create a React component that allows the user to select a color from a dropdown menu.Change the color of heading text to the selected color. The component should have an initial state value of "black" for the selected color. The dropdown menu should include at least four color options: black, red, blue, and green. Use the useState hook to manage the state of the selected color.