Q1
Create a HTML page with a heading and a paragraph.

Style the heading to have a font size of 36px and a font color of #333333.
Style the paragraph to have a font size of 18px, a font color of #666666, and have a line height of 1.5.
Add a margin of 20px between the heading and the paragraph.
Use default serif font-family for heading and paragraph.

Codesandbox : https://codesandbox.io/s/q1-css-ps1-jq87ki?file=/src/styles.css

Working Link : https://codesandbox.io/s/q1-css-ps1-jq87ki