Q4
Create a HTML page with two elements that include an input field and a submit button. Add CSS to style the elements as follows:

Style the input field to have a border of 1px solid #cccccc, a padding of 10px, and a font size of 18px.
Add a placeholder text of "Enter your email address" to the input field.
Style the submit button to have a background color of #333333, a font color of #ffffff, a font size of 18px and a padding of 10px 20px.
Add a margin of 10px between the input field and the submit button.

Codesandbox : https://codesandbox.io/s/q4-css-ps1-oydeyq?file=/index.html

Working Link : https://oydeyq.csb.app/