Q8
Adding on to the previous question, There should be three buttons for this purpose: "Low to High", "High to Low", and "Reset". When the user clicks on "Low to High", the products should be sorted by price in ascending order. When the user clicks on "High to Low", the products should be sorted by price in descending order. When the user clicks on "Reset", the products should be displayed in their original order.

Codesandbox : https://codesandbox.io/s/q8-ps7-90tokv?file=/src/App.jsx:0-1606

Working Link :  https://90tokv.csb.app/