Exercise 4
If a path segment starts with : then it becomes a "dynamic segment".
products/:productId - We call the productId in the URL "URL Params", or just "params" for short.

In a category listing page, on click of view details for each product, navigate to a page which shows the details of that particular product only. The URL should be unique for each product.

Codesandbox : https://codesandbox.io/s/q4-react-router-live-session-vbjyn4?file=/src/pages/phonesDB.js

Working Link : https://vbjyn4.csb.app/