Exercise 5
A <NavLink> is a special kind of <Link> that knows whether or not it is "active" or not.
This is useful when building a navigation menu, such as a breadcrumb or a set of tabs where you'd like to show which of them is currently selected.

Make the nav link on the top stand out when they're active i.e when that current page is active.

Codesandbox : https://codesandbox.io/s/q5-react-router-live-session-nn70lz?file=/src/App.js

Working Link : https://nn70lz.csb.app/