Q2,Q3
In the above question, in your product listing page add a “Add to Cart” button along with each product. On click of the button, add that item to your cart. The added product should be visible in the My Cart page. Show the total number of items available in the cart at the top of the page. Do this using context.

In the question above, add a “Remove from cart” button in the My Cart page for each product. On click of this button, the product should be removed from the cart

Codesandbox : https://codesandbox.io/s/q2-3-react-ps8-5mlfdi?file=/src/App.jsx

Working Link : https://5mlfdi.csb.app/