Q4,Q5
In the above question, in your product listing page add a “Add to Wishlist” button along with each product. On click of the button, add that item to your Wishlist page. The added product should be visible in the My Wishlist page. Show the total number of items available in the Wishlist at the top of the page. Do this using context.

In the question above, add a “Remove from Wishlist” button in the My Wishlist page for each product. On click of this button, the product should be removed from the Wishlist.

Codesandbox :  https://codesandbox.io/s/q4-5-react-ps8-u607d6?file=/src/App.jsx

Working Link : https://u607d6.csb.app/