ex02: create a store
The store is like a manager that keeps track of our app's state changes. It's responsible for making sure everything happens in the right order.

In our case, the store keeps track of our state, which is the number of cookies. We tell the store how our state should change using the reducer we made earlier.