ex04: dispatch actions
Dispatching actions is like telling our app what we want to do. It's like giving commands to our app to make things happen.

When we dispatch actions, we're telling our app what we want to do. We say, "Add a cookie" by dispatching an action that the store understands. The store then follows the plan with the help of reducer and changes the number of cookies accordingly.