ex08: displaying the todo list
Attaching the removeTodoHandler function to the window object allows it to be accessed from outside the current scope. In our case, it's likely that the function is being used as an event handler within the HTML content, which is not within the scope of the current JavaScript file.

By attaching it to the window object, you make sure that it can be called as a global function. This way, when the "Remove Todo" button is clicked in the HTML content, it can find and execute the removeTodoHandler function globally.

Codesandbox : https://codesandbox.io/s/rx1-2-cw-r2q7jk